organize theme hotspots components
