+++ src/app/hotspots/components/theme-top-header/theme-top-header.component.ts
@@ -1,15 +1,15 @@
 import { Component, ComponentFactoryResolver, ViewContainerRef, Input } from '@angular/core';
-import { PluginHotspot } from '../hotspot/plugin-hotspot';
-import { Profile } from '../models/profile';
+import { PluginHotspot } from '../../plugin-hotspot';
+import { Profile } from '../../../models/profile';
 
 @Component({
-  selector: 'app-theme-header',
+  selector: 'app-theme-top-header',
   template: ''
 })
-export class ThemeHeaderComponent extends PluginHotspot {
+export class ThemeTopHeaderComponent extends PluginHotspot {
 
   constructor(private viewContainerRef: ViewContainerRef, private factory: ComponentFactoryResolver, ) {
-    super('theme_header');
+    super('theme_top_header');
   }
 
   addHotspot(component: any) {
