+++ src/app/hotspots/hotspot.decorator.ts
