+++ src/app/hotspots/plugin-hotspot.ts
