+++ themes/brasil-que-o-povo-quer/index.ts
@@ -1,7 +1,7 @@
 import { HeaderComponent } from './header/header.component';
-import { ThemeMenuComponent } from './theme-menu/theme-menu.component';
+import { MenuComponent } from './menu/menu.component';
 import { FooterComponent } from './footer/footer.component';
 
 export let name = 'brasil-que-o-povo-quer';
 export let components: any = [];
-export let hotspots: any = [ HeaderComponent, ThemeMenuComponent, FooterComponent ];
+export let hotspots: any = [ HeaderComponent, MenuComponent, FooterComponent ];
