+++ themes/brasil-que-o-povo-quer/menu/menu.component.spec.ts
@@ -5,22 +5,22 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { By } from '@angular/platform-browser';
 
 import * as helpers from '../../../src/spec/helpers';
-import { ThemeMenuComponent } from './theme-menu.component';
+import { MenuComponent } from './menu.component';
 import { ProfileService } from '../../../src/app/services/profile.service';
 import { GlobalState } from '../../../src/app/global.state';
 import { Profile } from '../../../src/app/models/profile';
 import { CategoryService } from '../../../src/app/services/category.service';
 
-describe('ThemeMenuComponent', () => {
+describe('MenuComponent', () => {
 
-  let component: ThemeMenuComponent;
-  let fixture: ComponentFixture<ThemeMenuComponent>;
+  let component: MenuComponent;
+  let fixture: ComponentFixture<MenuComponent>;
   const mocks = helpers.getMocks();
 
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ThemeMenuComponent],
+      declarations: [MenuComponent],
       imports: [RouterTestingModule, TranslateModule.forRoot()],
       providers: [
         { provide: ProfileService, useValue: mocks.profileService },
@@ -33,7 +33,7 @@ describe('ThemeMenuComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ThemeMenuComponent);
+    fixture = TestBed.createComponent(MenuComponent);
     component = fixture.componentInstance;
     component.profile = <Profile>{ id: 1 };
     fixture.detectChanges();
