+++ themes/theme-default/footer/footer.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Hotspot } from '../../../src/app/hotspot/hotspot.decorator';
+import { Hotspot } from '../../../src/app/hotspots/hotspot.decorator';
 
 @Component({
   selector: 'theme-footer',
